Reorganize package
Make it so core is focused on projecting data onto an Alignment, ops contains generalized functions for geometries and graphs, and tools contains things to support the CLI.
