+++ surficial/core/api.py
@@ -1,5 +1,5 @@
 from surficial.core.alignment import Alignment
 from surficial.ops.graph import *
 from surficial.ops.shape import *
-from surficial.ops.calculations import *
-from surficial.ops.filter import *
+from surficial.core.calculations import *
+from surficial.core.filter import *
