+++ surficial/core/filter.py
