+++ surficial/tools/messages.py
@@ -1,5 +1,3 @@
 # Warnings
 UNPROJECTED = "{} is unprojected"
 PROJECTION = "CRS of {} differs from the CRS of the alignment {}"
-ISOLATED_NODES = "Found isolated nodes. Use the repair subcommand to check. Exiting now."
-MULTIPLE_SUBGRAPHS = "Found multiple subgraphs. Use the repaid subcommand to check. Exiting now."
