+++ surficial/tools/plotting.py
@@ -1,11 +1,20 @@
 from matplotlib.collections import LineCollection
 
 
-def cols_to_linecollection(vertices, xcol='x', ycol='y', style=None):
+def df_extents(df, xcol='x', ycol='y'):
+    from collections import namedtuple
+
+    Extents = namedtuple('Extents', ['minx', 'miny', 'maxx', 'maxy'])
+
+    extents = Extents(df[xcol].min(), df[ycol].min(), df[xcol].max(), df[ycol].max())
+    return extents
+
+
+def cols_to_linecollection(df, xcol='x', ycol='y', style=None):
     """Return a matplotlib LineCollection given two pandas DataFrame columns
 
     Parameters:
-        vertices (DataFrame): DataFrame containing xy data
+        df (DataFrame): DataFrame containing xy data
 
     Other Parameters:
         x (string): x column name
@@ -15,7 +24,7 @@ def cols_to_linecollection(vertices, xcol='x', ycol='y', style=None):
         collection (LineCollection): matplotlib LineCollection
 
     """
-    verts = [list(zip(edge[xcol], edge[ycol])) for _, edge in vertices.groupby('edge')]
+    verts = [list(zip(edge[xcol], edge[ycol])) for _, edge in df.groupby('edge')]
     if style != None:
         collection = LineCollection(verts, **style)
     else:
